Marrying various pieces together into detecting secrets flowing to a branch
The detection we have is a little too strict.  We miss one substantial case of false positives:

Secret determines branch, branches aren't both acyclic -- Bad and we catch it

Secret flow past branch, branches have different complexities not influenced
by the secret -- Ok, but we mark it as bad, thinking it's part of the above.

So, we'll have to do something to take care of the latter case eventually