Compute transitive closure of influence relationship Also, caught a bug where we thought constants were fresh variables.
There's still something unsatisifying about the computation of the transitive
closure relationship that I don't like.  We are definitely capturing flow, but
not at the right granularity. To wit, either it's important to capture the nature of the relationship (for purposes other than debugging) or we should be discarding the nature so that I can keep the sizes of the list of influencers down.