Finishing up rule weighting
Parser and ints disagree on the type of poly (3 * x vs. 3 * x + constant), so I
had to add a + Big_int.zero to each of the weights coming into the parser.  We
seem to get correct behavior on ingestion, but I know that the rule weights
aren't being used once we get them into the program.