Removes the CMeta path through koat
There were previously two major, nearly identical, paths from an input to an
answer in koat.  One across cprocs/cmeta, and the other across
cintprcos/cintmeta.  cmeta was used if all right-hand-sides were atoms.

In this case, the ITS was reduced to a TRS by calls to simple/simple.  Here, we
were losing information about weights on transitions.  Rather than shephearding
that information through, we decided to excise the second path.  cintmeta takes
a little more time, but produces tighter results in general (although both
processes agree on the order of the complexity of the two functions).