Changed how we pad arity
There are more restrictions oon the ITS / TRS than the functions having the same arity.
Turns out we need to have the same arguments in the same order on the left hand side,
so it's a little more complicated than just padding.

New apporach finds the max arity for the system and then produces a mapping from the argument list
[_x_0; _x_1; ...; _x_n] to the arguments of each function, and then extends as appropriate.