Fixes an issue where ruleCosts containing variables caused the complexity of a system to be unknown
In the final phases, when computing the total cost of a rule as being it's
transition cost times it's complexity (and those of the subrules), there was an
issue where rule costs were not in a particular special form (all variables of
the form X_i, where X_i relates to the i'th argument of a function).

A temporary fix is in place now, where we compute a substitution at the last
minute, and apply said substitution.