Arity fixing pushed into cint_aux
'parse' is deceptively named. It also massages the cint into a cannonical form.
When debugging some issues with aprove, it became apparent that we should be
doing the arity pads earlier on in the process.  Previously arity padding happened
after 'sanitize' was called.  Now we pad the arity out before our call to sanitize.

It's far easier to see that the translations being produced are faithful this way.