Split apart arity fixing and parsing, add a binary target that just fixes ITS system arity
One of the requirements of koat is that the functions all need to be of the same
arity.  We've previously added some functionality for translating cints (list of
comrule.ts) from whatever mixed arity configuration that they're in into a system
where all functions take the same number of arguments.

This adds a binary, fixArity, that takes a ITS and prints an equivalent ITS where all
functions have the same arity.  It's mostly for testing / debugging above functionality,
but might be useful on its own.