Bug Fix
Fixed a logic error. It's not the case that the rule that starts the loop (e.g. f0) will also
be the same as the index of that loop (as in, this is the 3rd loop).  This caused loop conditionals
to reference variables which didn't really exist, and made it look like the loops might not terminate.

We now carry around a little bit of extra state, the loop index, in the segment sum-type