Semantic Renaming
computeArity -> computeMinimumArity

This code is meant to generate test instances to debug some issues in slicing
and rule chaining. We aren't computing a single arity we intend to use in the
ITS that's passed on to koat.  We're computing a jumping off point from which
additional nonsense arguments can be added to produce slicable instances for
debugging.