Add generalised Its problem parser
Similar to the existing Cint parser. But uses a more simple and more general
parser for polynomials. Use flag `-use-its-parser` to activate it.

The Cint parser for polynomial expressions is complex and sometimes
fails to parse more complicated arithmetical expressions containing
parentheses and unary minus symbols. So it is necessary to provide a
"normalised" expression.

The new Its parser uses a more general (and more simple) expression
parser for polynomials. It has been compared against the Cint parser on
the koat-evaluation examples. No remarkable difference in performance
nor precision could be observed.